requirements.txt
streamlit>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
readability-lxml>=0.8.1
trafilatura>=1.6.0
python-dotenv>=1.0.0
validators>=0.22.0
google-genai>=1.0.0
ollama>=0.1.0
playwright>=1.40.0
openpyxl>=3.1.0
SQLAlchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
pandas>=2.0.0